Working on API Imp
Worked to send summary to API and then send it to the popup window, Also added a config file that stores our API key so its not hard coded into our JS files. Still getting a 429 error I believe its due to the fact we may be sending to much text to it.